* moved functions from main to lib to use it as a library
* removed a few unwraps in favour of Result inside the lib
