+++ src/main.rs
@@ -1,17 +1,5 @@
-mod constants;
-mod ds;
-mod process;
-
-use colored::*;
-use constants::Message;
-use ds::{key_node::KeyNode, mismatch::Mismatch};
-use serde_json;
-use std::{
-    fmt, fs,
-    io::{self, Write},
-    process as proc,
-    str::FromStr,
-};
+use json_diff::{compare_jsons, constants::Message, display_output};
+use std::{fmt, fs, process as proc, str::FromStr};
 use structopt::StructOpt;
 
 const HELP: &str = r#"
@@ -58,11 +46,6 @@ struct Cli {
     source2: String,
 }
 
-fn error_exit(message: constants::Message) -> ! {
-    eprintln!("{}", message);
-    proc::exit(1);
-}
-
 fn main() {
     let args = Cli::from_args();
 
@@ -80,198 +63,20 @@ fn main() {
             }
         }
     };
-    display_output(compare_jsons(&data1, &data2));
-}
-
-fn display_output(result: Mismatch) {
-    let no_mismatch = Mismatch {
-        left_only_keys: KeyNode::Nil,
-        right_only_keys: KeyNode::Nil,
-        keys_in_both: KeyNode::Nil,
+    let mismatch = match compare_jsons(&data1, &data2) {
+        Ok(mismatch) => mismatch,
+        Err(err) => {
+            eprintln!("{}", err);
+            proc::exit(1)
+        }
     };
-
-    let stdout = io::stdout();
-    let mut handle = io::BufWriter::new(stdout.lock());
-    if no_mismatch == result {
-        writeln!(handle, "\n{}", Message::NoMismatch).unwrap();
-    } else {
-        match result.keys_in_both {
-            KeyNode::Node(_) => {
-                let mut keys = Vec::new();
-                result.keys_in_both.absolute_keys(&mut keys, None);
-                writeln!(handle, "\n{}:", Message::Mismatch).unwrap();
-                for key in keys {
-                    writeln!(handle, "{}", key).unwrap();
-                }
-            }
-            KeyNode::Value(_, _) => writeln!(handle, "{}", Message::RootMismatch).unwrap(),
-            KeyNode::Nil => (),
-        }
-        match result.left_only_keys {
-            KeyNode::Node(_) => {
-                let mut keys = Vec::new();
-                result.left_only_keys.absolute_keys(&mut keys, None);
-                writeln!(handle, "\n{}:", Message::LeftExtra).unwrap();
-                for key in keys {
-                    writeln!(handle, "{}", key.red().bold()).unwrap();
-                }
-            }
-            KeyNode::Value(_, _) => error_exit(Message::UnknownError),
-            KeyNode::Nil => (),
-        }
-        match result.right_only_keys {
-            KeyNode::Node(_) => {
-                let mut keys = Vec::new();
-                result.right_only_keys.absolute_keys(&mut keys, None);
-                writeln!(handle, "\n{}:", Message::RightExtra).unwrap();
-                for key in keys {
-                    writeln!(handle, "{}", key.green().bold()).unwrap();
-                }
-            }
-            KeyNode::Value(_, _) => error_exit(Message::UnknownError),
-            KeyNode::Nil => (),
-        }
-    }
+    match display_output(mismatch) {
+        Ok(_) => (),
+        Err(err) => eprintln!("{}", err),
+    };
 }
 
-fn compare_jsons(a: &str, b: &str) -> Mismatch {
-    if let Ok(value1) = serde_json::from_str(a) {
-        if let Ok(value2) = serde_json::from_str(b) {
-            process::match_json(&value1, &value2)
-        } else {
-            error_exit(Message::JSON2);
-        }
-    } else {
-        error_exit(Message::JSON1);
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use maplit::hashmap;
-    use serde_json::json;
-
-    #[test]
-    fn nested_diff() {
-        let data1 = r#"{
-            "a":"b", 
-            "b":{
-                "c":{
-                    "d":true,
-                    "e":5,
-                    "f":9,
-                    "h":{
-                        "i":true,
-                        "j":false
-                    }
-                }
-            }
-        }"#;
-        let data2 = r#"{
-            "a":"b",
-            "b":{
-                "c":{
-                    "d":true,
-                    "e":6,
-                    "g":0,
-                    "h":{
-                        "i":false,
-                        "k":false
-                    }
-                }
-            }
-        }"#;
-
-        let expected_left = KeyNode::Node(hashmap! {
-        "b".to_string() => KeyNode::Node(hashmap! {
-                "c".to_string() => KeyNode::Node(hashmap! {
-                        "f".to_string() => KeyNode::Nil,
-                        "h".to_string() => KeyNode::Node( hashmap! {
-                                "j".to_string() => KeyNode::Nil,
-                            }
-                        ),
-                }
-                ),
-            }),
-        });
-        let expected_right = KeyNode::Node(hashmap! {
-            "b".to_string() => KeyNode::Node(hashmap! {
-                    "c".to_string() => KeyNode::Node(hashmap! {
-                            "g".to_string() => KeyNode::Nil,
-                            "h".to_string() => KeyNode::Node(hashmap! {
-                                    "k".to_string() => KeyNode::Nil,
-                                }
-                            )
-                        }
-                    )
-                }
-            )
-        });
-        let expected_uneq = KeyNode::Node(hashmap! {
-            "b".to_string() => KeyNode::Node(hashmap! {
-                    "c".to_string() => KeyNode::Node(hashmap! {
-                            "e".to_string() => KeyNode::Value(json!(5), json!(6)),
-                            "h".to_string() => KeyNode::Node(hashmap! {
-                                    "i".to_string() => KeyNode::Value(json!(true), json!(false)),
-                                }
-                            )
-                        }
-                    )
-                }
-            )
-        });
-        let expected = Mismatch::new(expected_left, expected_right, expected_uneq);
-
-        let mismatch = compare_jsons(data1, data2);
-        assert_eq!(mismatch, expected, "Diff was incorrect.");
-    }
-
-    #[test]
-    fn no_diff() {
-        let data1 = r#"{
-            "a":"b", 
-            "b":{
-                "c":{
-                    "d":true,
-                    "e":5,
-                    "f":9,
-                    "h":{
-                        "i":true,
-                        "j":false
-                    }
-                }
-            }
-        }"#;
-        let data2 = r#"{
-            "a":"b", 
-            "b":{
-                "c":{
-                    "d":true,
-                    "e":5,
-                    "f":9,
-                    "h":{
-                        "i":true,
-                        "j":false
-                    }
-                }
-            }
-        }"#;
-
-        assert_eq!(
-            compare_jsons(data1, data2),
-            Mismatch::new(KeyNode::Nil, KeyNode::Nil, KeyNode::Nil)
-        );
-    }
-
-    #[test]
-    fn no_json() {
-        let data1 = r#"{}"#;
-        let data2 = r#"{}"#;
-
-        assert_eq!(
-            compare_jsons(data1, data2),
-            Mismatch::new(KeyNode::Nil, KeyNode::Nil, KeyNode::Nil)
-        );
-    }
+fn error_exit(message: Message) -> ! {
+    eprintln!("{}", message);
+    proc::exit(1);
 }
